Additional Testing and modifying Foreign Key relations on Migrations

+++ database/migrations/2019_10_03_120312_create_topics_table.php
@@ -17,8 +17,8 @@ class CreateTopicsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name')->comment('Project Name');
             $table->longText('description')->comment('Project Description');
-            $table->unsignedInteger('studentID')->comment('Foreign Key for STUDENTS')->nullable();
-            $table->unsignedInteger('supervisorID')->comment('Foreign Key for SUPERVISORS');
+            $table->unsignedBigInteger('studentID')->comment('Foreign Key for STUDENTS')->nullable();
+            $table->unsignedBigInteger('supervisorID')->comment('Foreign Key for SUPERVISORS');
             $table->boolean('isMCApprove')->comment('Boolean: CS Multimedia Approved');
             $table->boolean('isCBApprove')->comment('Boolean: CS Business Approved');
